requirements.txt
colorama
requests
dnspython
random-header-generator
bs4